+++ src/main/java/blockteil/filters/SmallThresholdFilterer.java
@@ -17,10 +17,10 @@ public class SmallThresholdFilterer extends KMERFilterer {
             if (Config.OFFSET == 1 && i > 0) kmer = revcomp ? KMER.shiftKMER_revcomp(kmer, seq.charAt(i + Config.KMER_LENGTH - 1)) : KMER.shiftKMER(kmer, seq.charAt(i + Config.KMER_LENGTH - 1));
             else kmer = revcomp ? KMER.makeKMER_revcomp(seq, i) : KMER.makeKMER(seq, i);
 
-            Set<Integer> matchingGenes = Config.KMER_MAP.get(kmer);
+            Set<Short> matchingGenes = Config.KMER_MAP.get(kmer);
             if (matchingGenes == null) continue;
 
-            for (Integer gene : matchingGenes) {
+            for (short gene : matchingGenes) {
                 out.set(gene);
                 if (Config.EARLY_TERMINATION_ALLOWED) break A;
             }
